pkg/adaptor: Call PullImage before CreateContainer
The container runtime implementations such as containerd and cri-o
do not call PullImage of kata CCv0 to pull images in pod VMs. So far,
we need a custom version of containerd or kata shim to pull images
in pod VMs.

With this patch, cloud-api-adaptor calls PullImage to pull an image
in a pod VM, while processing a CreateContainer request.

Fixes #88

Signed-off-by: Yohei Ueda <yohei@jp.ibm.com>