podvm: rhel: relabel before running misc-settings.sh
The relabeling in the provisioner is intended to fix labeling when
packer copies over files to the image. It's run at the end during
provisioning. However, misc-settings.sh may install packages and
configure services to boot at startup. If the labels are not correct,
this will fail.

Move labeling to right after we copy our target dir and run misc-settings
in the end.

Fixes: #548
Signed-off-by: Bandan Das <bsd@redhat.com>