test/e2e: add step to upload the podvm image
This added into the e2e workflow a step about uploading the podvm qcow2
image into the VPC image storage. The step is triggered only when the
variable TEST_E2E_PODVM_IMAGE, which contains the path to the qcow2
file, is exported.

Each and every implementer of the CloudProvision interface must
implement the UploadPodvm() function. This provides an implementation
for the LibvirtProvisioner.

Fixes #592
Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>