test: Fix undeclared name error
The testEnv variable is defined in main_test.go,
but is used in a function in common_suite.go,
which is not a form of *_test.go.

This function is only called from a test, but IDEs
like VS Code do not recognize the situation,
and show an undeclared name error.

This patch renames common_suite.go to
common_suite_test.go to suppress the error.

Fixes #670

Signed-off-by: Yohei Ueda <yohei@jp.ibm.com>