podvm: Disable VNC of Packer QEMU Builder
Currently, the boot_wait option is specified to packer. This option is
necessary, because boot_command is specified and sent to the started VM
via VNC. We are using public cloud VM images, and we do not need to send
boot commands via VNC to start up VM.

This patch removes the boot_command option, and adds the disable_vnc
option. This change makes packer wait using SSH instead of VNC for VM
startup.

Fixes #981

Signed-off-by: Yohei Ueda <yohei@jp.ibm.com>