pkg/azure: Delete resources with VM
By using the DeleteOption when creating the VM, we can ask the Azure Resource
Manager to cleanup disk and NIC when the VM is deleted. This allows us to remove
the cleanup code from DeleteInstance(). The only place where we need the cleanup
code is when creating the instance fails.

Fixes: #1203
Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>