packer: Enable misc addons for podvm
This commit enables configuring different addons for aws and azure
podvm image. For example enabling gpus, blobfuse etc.

There is specific directory and file structure used to make
it generic - podvm/addons/{addon-name}

Each addon-name needs to have setup.sh for setting up the addon.
Any configs or binaries for the addon need to be in the respective
{addon-name} folder.

To enable the addons create a file named .enable in the top level podvm/addons
directory and any of the {addon-name} subdir that you want to include.

Then you can build the aws or azure image using packer.

For example:
cd azure/image
PODVM_DISTRO=ubuntu make image

The specific addons are setup by the qcow2/setup_addons.sh script

A skeleton addon folder is provided as well to help with understanding
the directory structure

Signed-off-by: Pradipta Banerjee <pradipta.banerjee@gmail.com>