cloud: Avoid deletion of instance for debugging
Provide an env variable KEEP_INSTANCE, which if set to "true"
will ensure the VM is not deleted on any error. This helps
with debugging.
Further, to use this functionality for debugging, it's better
to create a pod with `restartPolicy: Never` in the spec
Otherwise new VMs will keep getting created

Signed-off-by: Pradipta Banerjee <pradipta.banerjee@gmail.com>